LayoutTests/fast/events/click-over-descendant-elements-expected.txt

Check if any mousedown-mouseup pairs in one element dispatch click events.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Very normal click on container:
PASS lastClickTarget is container
Move inside container:
PASS lastClickTarget is container
Move from container to a child:
PASS lastClickTarget is container
Move from a child to container:
PASS lastClickTarget is container
Move from a child to another child:
PASS lastClickTarget is container
Move out from a child:
PASS lastClickTarget is document.body
Removing a child element in the light DOM during a click:
FAIL lastClickTarget should be null. Was [object HTMLDivElement].
Click on escaping button content:
PASS lastClickTarget is button1
Click on disappearing INPUT value:
PASS lastClickTarget is input1
PASS successfullyParsed is true

TEST COMPLETE
